Conflicts:
	cocos/2d/CCActionGrid.h
	cocos/2d/CCActionGrid3D.h
	cocos/2d/CCActionInstant.h
	cocos/2d/CCActionInterval.h
	cocos/2d/CCActionPageTurn3D.h
	cocos/2d/CCActionProgressTimer.h
	cocos/2d/CCActionTiledGrid.h
	cocos/2d/CCActionTween.h
	cocos/2d/CCAnimation.h
	cocos/2d/CCAutoPolygon.h
	cocos/2d/CCCamera.h
	cocos/2d/CCClippingNode.h
	cocos/2d/CCClippingRectangleNode.h
	cocos/2d/CCDrawNode.h
	cocos/2d/CCFastTMXLayer.h
	cocos/2d/CCGrid.h
	cocos/2d/CCLabel.h
	cocos/2d/CCLayer.h
	cocos/2d/CCMenuItem.h
	cocos/2d/CCNode.h
	cocos/2d/CCParallaxNode.h
	cocos/2d/CCParticleSystem.h
	cocos/2d/CCParticleSystemQuad.h
	cocos/2d/CCProtectedNode.h
	cocos/2d/CCRenderTexture.h
	cocos/2d/CCSprite.h
	cocos/2d/CCSpriteFrameCache.h
	cocos/2d/CCTMXLayer.h
	cocos/2d/CCTMXTiledMap.h
	cocos/2d/CCTextFieldTTF.h
	cocos/3d/CCAABB.h
	cocos/3d/CCAnimate3D.h
	cocos/3d/CCAnimationCurve.h
	cocos/3d/CCBundleReader.h
	cocos/3d/CCMesh.h
	cocos/3d/CCOBB.h
	cocos/3d/CCSkeleton3D.h
	cocos/3d/CCSkybox.h
	cocos/3d/CCSprite3D.h
	cocos/3d/CCSprite3DMaterial.h
	cocos/3d/CCTerrain.h
	cocos/audio/include/AudioEngine.h
	cocos/base/CCAutoreleasePool.h
	cocos/base/CCData.h
	cocos/base/CCDirector.h
	cocos/base/CCEvent.h
	cocos/base/CCEventController.h
	cocos/base/CCEventDispatcher.h
	cocos/base/CCEventListener.h
	cocos/base/CCIMEDispatcher.h
	cocos/base/CCProtocols.h
	cocos/base/CCRef.h
	cocos/base/CCScriptSupport.h
	cocos/base/CCTouch.h
	cocos/base/CCVector.h
	cocos/base/ccConfig.h
	cocos/editor-support/cocostudio/ActionTimeline/CCActionTimeline.h
	cocos/editor-support/cocostudio/CCActionFrame.h
	cocos/editor-support/cocostudio/CCActionManagerEx.h
	cocos/editor-support/cocostudio/CCActionNode.h
	cocos/editor-support/cocostudio/CCArmatureAnimation.h
	cocos/editor-support/cocostudio/CCArmatureDataManager.h
	cocos/editor-support/cocostudio/CCComAttribute.h
	cocos/editor-support/cocostudio/CCComAudio.h
	cocos/editor-support/cocostudio/CCComController.h
	cocos/editor-support/cocostudio/CCComRender.h
	cocos/math/CCAffineTransform.h
	cocos/math/Mat4.h
	cocos/network/CCDownloader.h
	cocos/network/HttpClient.h
	cocos/network/HttpRequest.h
	cocos/network/HttpResponse.h
	cocos/network/SocketIO.h
	cocos/network/WebSocket.h
	cocos/physics/CCPhysicsBody.h
	cocos/physics/CCPhysicsContact.h
	cocos/physics/CCPhysicsHelper.h
	cocos/physics/CCPhysicsJoint.h
	cocos/physics/CCPhysicsShape.h
	cocos/physics/CCPhysicsWorld.h
	cocos/physics3d/CCPhysics3DConstraint.h
	cocos/physics3d/CCPhysics3DObject.h
	cocos/physics3d/CCPhysics3DShape.h
	cocos/platform/CCFileUtils.h
	cocos/platform/CCImage.h
	cocos/renderer/CCGLProgramState.h
	cocos/renderer/CCGroupCommand.h
	cocos/renderer/CCMaterial.h
	cocos/renderer/CCPrimitive.h
	cocos/renderer/CCPrimitiveCommand.h
	cocos/renderer/CCRenderCommand.h
	cocos/renderer/CCRenderState.h
	cocos/renderer/CCRenderer.h
	cocos/renderer/CCTexture2D.h
	cocos/renderer/CCTextureCache.h
	cocos/renderer/CCTrianglesCommand.h
	cocos/renderer/CCVertexIndexBuffer.h
	cocos/renderer/CCVertexIndexData.h
	cocos/scripting/js-bindings/manual/ScriptingCore.h
	cocos/scripting/lua-bindings/manual/CCLuaBridge.h
	cocos/scripting/lua-bindings/manual/CCLuaEngine.h
	cocos/scripting/lua-bindings/manual/CCLuaStack.h
	cocos/scripting/lua-bindings/manual/CCLuaValue.h
	cocos/scripting/lua-bindings/manual/LuaBasicConversions.h
	cocos/scripting/lua-bindings/manual/cocos2d/LuaOpengl.h
	cocos/scripting/lua-bindings/manual/cocos2d/LuaScriptHandlerMgr.h
	cocos/scripting/lua-bindings/manual/cocosbuilder/CCBProxy.h
	cocos/scripting/lua-bindings/manual/platform/android/CCLuaJavaBridge.h
	cocos/scripting/lua-bindings/manual/tolua_fix.h
	cocos/ui/UIEditBox/UIEditBox.h
	cocos/ui/UIHelper.h
	cocos/ui/UILayout.h
	cocos/ui/UILayoutComponent.h
	cocos/ui/UIListView.h
	cocos/ui/UIPageView.h
	cocos/ui/UIScale9Sprite.h
	cocos/ui/UIScrollView.h
	cocos/ui/UIText.h
	cocos/ui/UITextField.h
	cocos/ui/UIVideoPlayer.h
	cocos/ui/UIWebView.h
	cocos/ui/UIWidget.h
	extensions/assets-manager/AssetsManager.h
	extensions/assets-manager/AssetsManagerEx.h
	extensions/assets-manager/Manifest.h
	plugin
	tests/cpp-tests/Resources/ccs-res
	tools/bindings-generator
	tools/cocos2d-console
	web

+++ extensions/assets-manager/AssetsManagerEx.cpp
@@ -33,6 +33,7 @@
 #else // from our embedded sources
 #include "unzip.h"
 #endif
+#include "base/CCAsyncTaskPool.h"
 
 using namespace cocos2d;
 using namespace std;
@@ -50,7 +51,6 @@ NS_CC_EXT_BEGIN
 
 const std::string AssetsManagerEx::VERSION_ID = "@version";
 const std::string AssetsManagerEx::MANIFEST_ID = "@manifest";
-const std::string AssetsManagerEx::BATCH_UPDATE_ID = "@batch_update";
 
 // Implementation of AssetsManagerEx
 
@@ -80,16 +80,19 @@ AssetsManagerEx::AssetsManagerEx(const std::string& manifestUrl, const std::stri
     _updateState = State::UNCHECKED;
 
     _downloader = std::shared_ptr<network::Downloader>(new network::Downloader);
-    _downloader->setConnectionTimeout(DEFAULT_CONNECTION_TIMEOUT);
-    _downloader->setErrorCallback(std::bind(&AssetsManagerEx::onError, this, std::placeholders::_1));
-    _downloader->setProgressCallback(std::bind(&AssetsManagerEx::onProgress,
-                                         this,
-                                         std::placeholders::_1,
-                                         std::placeholders::_2,
-                                         std::placeholders::_3,
-                                         std::placeholders::_4)
-                                     );
-    _downloader->setSuccessCallback(std::bind(&AssetsManagerEx::onSuccess, this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+//    _downloader->setConnectionTimeout(DEFAULT_CONNECTION_TIMEOUT);
+    _downloader->onTaskError = bind(&AssetsManagerEx::onError, this, placeholders::_1, placeholders::_2, placeholders::_3, placeholders::_4);
+    _downloader->onTaskProgress = [this](const network::DownloadTask& task,
+                                         int64_t bytesReceived,
+                                         int64_t totalBytesReceived,
+                                         int64_t totalBytesExpected)
+    {
+        this->onProgress(totalBytesExpected, totalBytesReceived, task.requestURL, task.identifier);
+    };
+    _downloader->onFileTaskSuccess = [this](const network::DownloadTask& task)
+    {
+        this->onSuccess(task.requestURL, task.storagePath, task.identifier);
+    };
     setStoragePath(storagePath);
     _cacheVersionPath = _storagePath + VERSION_FILENAME;
     _cacheManifestPath = _storagePath + MANIFEST_FILENAME;
@@ -100,9 +103,9 @@ AssetsManagerEx::AssetsManagerEx(const std::string& manifestUrl, const std::stri
 
 AssetsManagerEx::~AssetsManagerEx()
 {
-    _downloader->setErrorCallback(nullptr);
-    _downloader->setSuccessCallback(nullptr);
-    _downloader->setProgressCallback(nullptr);
+    _downloader->onTaskError = (nullptr);
+    _downloader->onFileTaskSuccess = (nullptr);
+    _downloader->onTaskProgress = (nullptr);
     CC_SAFE_RELEASE(_localManifest);
     // _tempManifest could share a ptr with _remoteManifest or _localManifest
     if (_tempManifest != _localManifest && _tempManifest != _remoteManifest)
@@ -138,7 +141,7 @@ void AssetsManagerEx::initManifests(const std::string& manifestUrl)
         if (_tempManifest)
         {
             _tempManifest->parse(_tempManifestPath);
-            if (!_tempManifest->isLoaded())
+            if (!_tempManifest->isLoaded() && _fileUtils->isFileExist(_tempManifestPath))
                 _fileUtils->removeFile(_tempManifestPath);
         }
         else
@@ -288,7 +291,7 @@ bool AssetsManagerEx::decompress(const std::string &zip)
     const std::string rootPath = zip.substr(0, pos+1);
     
     // Open the zip file
-    unzFile zipfile = unzOpen(zip.c_str());
+    unzFile zipfile = unzOpen(FileUtils::getInstance()->getSuitableFOpen(zip).c_str());
     if (! zipfile)
     {
         CCLOG("AssetsManagerEx : can not open downloaded zip file %s\n", zip.c_str());
@@ -440,7 +443,7 @@ void AssetsManagerEx::downloadVersion()
     {
         _updateState = State::DOWNLOADING_VERSION;
         // Download version file asynchronously
-        _downloader->downloadAsync(versionUrl, _cacheVersionPath, VERSION_ID);
+        _downloader->createDownloadFileTask(versionUrl, _cacheVersionPath, VERSION_ID);
     }
     // No version file found
     else
@@ -496,7 +499,7 @@ void AssetsManagerEx::downloadManifest()
     {
         _updateState = State::DOWNLOADING_MANIFEST;
         // Download version file asynchronously
-        _downloader->downloadAsync(manifestUrl, _tempManifestPath, MANIFEST_ID);
+        _downloader->createDownloadFileTask(manifestUrl, _tempManifestPath, MANIFEST_ID);
     }
     // No manifest file found
     else
@@ -559,9 +562,8 @@ void AssetsManagerEx::startUpdate()
     if (_tempManifest->isLoaded() && _tempManifest->versionEquals(_remoteManifest))
     {
         _tempManifest->genResumeAssetsList(&_downloadUnits);
-        
         _totalWaitToDownload = _totalToDownload = (int)_downloadUnits.size();
-        _downloader->batchDownloadAsync(_downloadUnits, BATCH_UPDATE_ID);
+        this->batchDownload();
         
         std::string msg = StringUtils::format("Resuming from previous unfinished update, %d files remains to be finished.", _totalToDownload);
         dispatchUpdateEvent(EventAssetsManagerEx::EventCode::UPDATE_PROGRESSION, "", msg);
@@ -598,16 +600,15 @@ void AssetsManagerEx::startUpdate()
                     // Create path
                     _fileUtils->createDirectory(basename(_storagePath + path));
 
-                    network::DownloadUnit unit;
+                    DownloadUnit unit;
                     unit.customId = it->first;
                     unit.srcUrl = packageUrl + path;
                     unit.storagePath = _storagePath + path;
-                    unit.resumeDownload = false;
                     _downloadUnits.emplace(unit.customId, unit);
                 }
             }
             // Set other assets' downloadState to SUCCESSED
-            auto assets = _remoteManifest->getAssets();
+            auto &assets = _remoteManifest->getAssets();
             for (auto it = assets.cbegin(); it != assets.cend(); ++it)
             {
                 const std::string &key = it->first;
@@ -617,9 +618,8 @@ void AssetsManagerEx::startUpdate()
                     _tempManifest->setAssetDownloadState(key, Manifest::DownloadState::SUCCESSED);
                 }
             }
-            
             _totalWaitToDownload = _totalToDownload = (int)_downloadUnits.size();
-            _downloader->batchDownloadAsync(_downloadUnits, BATCH_UPDATE_ID);
+            this->batchDownload();
             
             std::string msg = StringUtils::format("Start to update %d files from remote package.", _totalToDownload);
             dispatchUpdateEvent(EventAssetsManagerEx::EventCode::UPDATE_PROGRESSION, "", msg);
@@ -641,12 +641,50 @@ void AssetsManagerEx::updateSucceed()
     _remoteManifest = nullptr;
     // 3. make local manifest take effect
     prepareLocalManifest();
+
+
+    _updateState = State::UNZIPPING;
     // 4. decompress all compressed files
-    decompressDownloadedZip();
-    // 5. Set update state
-    _updateState = State::UP_TO_DATE;
-    // 6. Notify finished event
-    dispatchUpdateEvent(EventAssetsManagerEx::EventCode::UPDATE_FINISHED);
+    //decompressDownloadedZip();
+
+    struct AsyncData
+    {
+        std::vector<std::string> compressedFiles;
+        std::string errorCompressedFile;
+    };
+
+    AsyncData* asyncData = new AsyncData;
+    asyncData->compressedFiles = _compressedFiles;
+    _compressedFiles.clear();
+
+    std::function<void(void*)> mainThread = [this](void* param) {
+        AsyncData* asyncData = (AsyncData*)param;
+        if (asyncData->errorCompressedFile.empty())
+        {
+            // 5. Set update state
+            _updateState = State::UP_TO_DATE;
+            // 6. Notify finished event
+            dispatchUpdateEvent(EventAssetsManagerEx::EventCode::UPDATE_FINISHED);
+        }
+        else
+        {
+            _updateState = State::FAIL_TO_UPDATE;
+            dispatchUpdateEvent(EventAssetsManagerEx::EventCode::ERROR_DECOMPRESS, "", "Unable to decompress file " + asyncData->errorCompressedFile);
+        }
+
+        delete asyncData;
+    };
+    AsyncTaskPool::getInstance()->enqueue(AsyncTaskPool::TaskType::TASK_OTHER, mainThread, (void*)asyncData, [this, asyncData]() {
+        // Decompress all compressed files
+        for (auto& zipFile : asyncData->compressedFiles) {
+            if (!decompress(zipFile))
+            {
+                asyncData->errorCompressedFile = zipFile;
+                break;
+            }
+            _fileUtils->removeFile(zipFile);
+        }
+    });
 }
 
 void AssetsManagerEx::checkUpdate()
@@ -745,6 +783,7 @@ void AssetsManagerEx::update()
             break;
         case State::UP_TO_DATE:
         case State::UPDATING:
+        case State::UNZIPPING:
             _waitToUpdate = false;
             break;
         default:
@@ -752,7 +791,7 @@ void AssetsManagerEx::update()
     }
 }
 
-void AssetsManagerEx::updateAssets(const network::DownloadUnits& assets)
+void AssetsManagerEx::updateAssets(const DownloadUnits& assets)
 {
     if (!_inited){
         CCLOG("AssetsManagerEx : Manifests uninited.\n");
@@ -768,7 +807,7 @@ void AssetsManagerEx::updateAssets(const network::DownloadUnits& assets)
             _updateState = State::UPDATING;
             _downloadUnits.clear();
             _downloadUnits = assets;
-            _downloader->batchDownloadAsync(_downloadUnits, BATCH_UPDATE_ID);
+            this->batchDownload();
         }
         else if (size == 0 && _totalWaitToDownload == 0)
         {
@@ -777,7 +816,7 @@ void AssetsManagerEx::updateAssets(const network::DownloadUnits& assets)
     }
 }
 
-const network::DownloadUnits& AssetsManagerEx::getFailedAssets() const
+const DownloadUnits& AssetsManagerEx::getFailedAssets() const
 {
     return _failedUnits;
 }
@@ -789,29 +828,32 @@ void AssetsManagerEx::downloadFailedAssets()
 }
 
 
-void AssetsManagerEx::onError(const network::Downloader::Error &error)
+void AssetsManagerEx::onError(const network::DownloadTask& task,
+                              int errorCode,
+                              int errorCodeInternal,
+                              const std::string& errorStr)
 {
     // Skip version error occured
-    if (error.customId == VERSION_ID)
+    if (task.identifier == VERSION_ID)
     {
         CCLOG("AssetsManagerEx : Fail to download version file, step skipped\n");
         _updateState = State::PREDOWNLOAD_MANIFEST;
         downloadManifest();
     }
-    else if (error.customId == MANIFEST_ID)
+    else if (task.identifier == MANIFEST_ID)
     {
-        dispatchUpdateEvent(EventAssetsManagerEx::EventCode::ERROR_DOWNLOAD_MANIFEST, error.customId, error.message, error.curle_code, error.curlm_code);
+        dispatchUpdateEvent(EventAssetsManagerEx::EventCode::ERROR_DOWNLOAD_MANIFEST, task.identifier, errorStr, errorCode, errorCodeInternal);
     }
     else
     {
-        auto unitIt = _downloadUnits.find(error.customId);
+        auto unitIt = _downloadUnits.find(task.identifier);
         // Found unit and add it to failed units
         if (unitIt != _downloadUnits.end())
         {
-            network::DownloadUnit unit = unitIt->second;
+            DownloadUnit unit = unitIt->second;
             _failedUnits.emplace(unit.customId, unit);
         }
-        dispatchUpdateEvent(EventAssetsManagerEx::EventCode::ERROR_UPDATING, error.customId, error.message, error.curle_code, error.curlm_code);
+        dispatchUpdateEvent(EventAssetsManagerEx::EventCode::ERROR_UPDATING, task.identifier, errorStr, errorCode, errorCodeInternal);
     }
 }
 
@@ -879,27 +921,9 @@ void AssetsManagerEx::onSuccess(const std::string &srcUrl, const std::string &st
         _updateState = State::MANIFEST_LOADED;
         parseManifest();
     }
-    else if (customId == BATCH_UPDATE_ID)
-    {
-        // Finished with error check
-        if (_failedUnits.size() > 0 || _totalWaitToDownload > 0)
-        {
-            // Save current download manifest information for resuming
-            _tempManifest->saveToFile(_tempManifestPath);
-            
-            decompressDownloadedZip();
-            
-            _updateState = State::FAIL_TO_UPDATE;
-            dispatchUpdateEvent(EventAssetsManagerEx::EventCode::UPDATE_FAILED);
-        }
-        else
-        {
-            updateSucceed();
-        }
-    }
     else
     {
-        auto assets = _remoteManifest->getAssets();
+        auto &assets = _remoteManifest->getAssets();
         auto assetIt = assets.find(customId);
         if (assetIt != assets.end())
         {
@@ -932,6 +956,25 @@ void AssetsManagerEx::onSuccess(const std::string &srcUrl, const std::string &st
             // Remove from failed units list
             _failedUnits.erase(unitIt);
         }
+        
+        if (_totalWaitToDownload <= 0)
+        {
+            // Finished with error check
+            if (_failedUnits.size() > 0)
+            {
+                // Save current download manifest information for resuming
+                _tempManifest->saveToFile(_tempManifestPath);
+                
+                decompressDownloadedZip();
+                
+                _updateState = State::FAIL_TO_UPDATE;
+                dispatchUpdateEvent(EventAssetsManagerEx::EventCode::UPDATE_FAILED);
+            }
+            else
+            {
+                updateSucceed();
+            }
+        }
     }
 }
 
@@ -939,6 +982,15 @@ void AssetsManagerEx::destroyDownloadedVersion()
 {
     _fileUtils->removeFile(_cacheVersionPath);
     _fileUtils->removeFile(_cacheManifestPath);
+}
+
+void AssetsManagerEx::batchDownload()
+{
+    for(auto iter : _downloadUnits)
+    {
+        DownloadUnit& unit = iter.second;
+        _downloader->createDownloadFileTask(unit.srcUrl, unit.storagePath, unit.customId);
+    }
 }
 
 NS_CC_EXT_END
